app/databricks_app/requirements.txt
pyspark
matplotlib
ipywidgets
sentence-transformers
psycopg2-binary
azure-identity
azure-keyvault-secrets
numpy

